feat: Update ML classifier and frontend components for enhanced category prediction and analysis

+++ backend/spring-boot/src/main/java/com/ceview/module1/businessinput/dto/AnalyzeDtos.java
@@ -12,7 +12,7 @@ public class AnalyzeDtos {
         String uvp
     ) {}
 
-    public record CategoryAllocation(String name, double percentage) {}
+    public record CategoryAllocation(String name, int percentage) {}
 
     public record AnalyzeResponse(List<CategoryAllocation> categories) {}
